fix: move header componeent out of the home folder it is unrelated to ticket
