+++ src/layout/index.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import Header from '../components/Home/Header';
+import Header from '../components/Header';
 import Footer from './Footer';
 
 const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
